+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 # 0.2.8
 
 1. Rename method `FormService::addFormOptions()` -> `FormService::addHtml5ValidationOptions()`
+2. Fix failed tests of the `Descriptors` class (collection of `Descriptor`)
 
 # 0.2.7